Refactor contents of components/utils to better reflect contents
